requirements.txt
python-dotenv==1.1.0
aiogram==3.20.0.post0
aiosqlite==0.21.0
asyncpg==0.30.0